pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-llm-developer"
version = "0.1.0"
description = "Local-first AI software-engineering orchestration platform for consumer GPUs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "lstred" }]
keywords = ["llm", "agents", "ollama", "llama.cpp", "code-generation", "local-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.7",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "tenacity>=8.2",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]
llamacpp = [
    "llama-cpp-python>=0.2.77",
]

[project.scripts]
local-llm-dev = "lld.cli:app"

[project.urls]
Homepage = "https://github.com/lstred/local-llm-developer"
Repository = "https://github.com/lstred/local-llm-developer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"lld" = ["prompts/*.md", "prompts/**/*.md", "ui/static/*", "ui/templates/*.html"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.10"
strict_optional = true
warn_unused_ignores = true
ignore_missing_imports = true
